Module refactor: break down code from plugins/vt.py
Into modules that resemble the Avocado core structure, and will
hopefully make code easier to grasp and maintain.
